% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addClusters.R
\name{addClusters}
\alias{addClusters}
\title{Clustering cells}
\usage{
addClusters(
  mae,
  name.iterative.lsi = "iterativeLSI",
  clusters.colname = "Clusters",
  method = "Seurat",
  force = FALSE
)
}
\arguments{
\item{mae}{MultiAssayExperiment}

\item{name.iterative.lsi}{String specifying the name of the reduced dimensions to create clusters from}

\item{clusters.colname}{String specifying the column name to save the clusters as in the experiment column data}

\item{method}{String specifying the method for creating clsuters. Valid options are "Seurat" or "Scran".}

\item{force}{Logical whether to overwrite existing columns with clusters.colname column name}
}
\value{
A \linkS4class{MultiAssayExperiment}
}
\description{
Creating cell clusters from iterative LSI reduced dimensions contained with a MAE and adding the cluster result back to the MAE
}
\author{
Natalie Fox
}
