% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDoubletScores.R
\name{addDoubletScores}
\alias{addDoubletScores}
\title{Add doublet scores}
\usage{
addDoubletScores(
  mae,
  experiment.name = "TileMatrix500",
  seed = 5,
  num.trials = 5,
  num.threads = 1,
  num.dimensions = 30,
  plot.out.dir = ".",
  BPPARAM = bpparam()
)
}
\arguments{
\item{mae}{\linkS4class{MultiAssayExperiment}.}

\item{experiment.name}{String containing thr experiment name to calculate doublet scores.}

\item{seed}{Numeric scalar to be used as the seed for random number generation. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{num.trials}{Integer scalar specifying the number of times to simulate the number of cells in the sample for doublets.}

\item{num.threads}{Integer scalar specifying the number of threads to be used for parallel computing.}

\item{num.dimensions}{Integer scalar specifying the number of iterative LSI dimensions to use doublet scoring}

\item{plot.out.dir}{String spcifying the output path for doublet score quality control plots}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object indicating how doublet scoring should be parallelized per sample.}
}
\value{
\linkS4class{MultiAssayExperiment} with doublet scores column added to the experiment colData.
}
\description{
Calculate doublet scores and add them to the MultiAssayExperiment.
}
\author{
Natalie Fox
}
