% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeneIntegrationMatrix.R
\name{addGeneIntegrationMatrix}
\alias{addGeneIntegrationMatrix}
\title{Add gene integration matrix}
\usage{
addGeneIntegrationMatrix(
  mae,
  rna.sce = NULL,
  gene.score.experiment.name = "GeneScoreMatrix",
  gene.expression.experiment.name = "GeneExpressionMatrix",
  name.reduced.dim = c(TileMatrix500 = "iterativeLSI"),
  group.gex = c(GeneExpressionMatrix = "Clusters"),
  group.list = NULL,
  sampleCellsATAC = 10000,
  sampleCellsRNA = 10000,
  reduction = "cca",
  genesUse = NULL,
  num.gene.features = 2000
)
}
\arguments{
\item{mae}{\linkS4class{MultiAssayExperiment} containing scATAC-seq gene score experiment and optionally the scRNA-seq experiment.}

\item{rna.sce}{\linkS4class{SingleCellExperiment} or \linkS4class{RangedSummarizedExperiment} or \linkS4class{RangedSummarizedExperiment} specifying scRNA-seq sata to integrated with the scATAC-seq data. This can be NULL if the RNA experiment is already in the mae.}

\item{gene.score.experiment.name}{String containing the experiment name of the gene score matrix calculated from ATAC data.}

\item{gene.expression.experiment.name}{String containing the experiment name of gene expression matrix (RNA data).}

\item{name.reduced.dim}{String containing the reduced dimension name.}

\item{group.gex}{String specifying the column name for RNA groups. If the column is not in the gene.expression.experiment.name matrix, the specifying the correct experiment name as the name in a named vector.}

\item{group.list}{A list of strings specifying the cells to be used for RNA-ATAC integration. This is used to constrain integration across biologically relevant groups. Format of the list must be subgroups with RNA and ATAC sets. For example: list(groupA = list(ATAC = cellsATAC_A, RNA = cellsRNA_A), groupB = list(ATAC = cellsATAC_B, RNA = cellsRNA_B))}

\item{sampleCellsATAC}{Scalar integer specifying the number of scATAC-seq cells to use for integration.}

\item{sampleCellsRNA}{Scalar integer specifying the number of scRNA-seq cells to use for integration.}

\item{reduction}{String specifying the seurat method to use for integrating modalities. See Seuat::FindTransferAnchors() for details.}

\item{genesUse}{String vector specifying gene names to use for integration.}

\item{num.gene.features}{Scalar integer specifying the number of variable genes determined by `Seurat::FindVariableGenes()`}
}
\value{
\linkS4class{MultiAssayExperiment} with DoubletScore and DoubletEnrichment columns added to the experiment colData.
}
\description{
Calculate gene integration matrix and add the matrix to the MultiAssayExperiment.
}
\author{
Natalie Fox
}
