% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIterativeLSI.R
\name{addIterativeLSI}
\alias{addIterativeLSI}
\title{Add new iterative LSI embeddings to MultiAssayExperiment}
\usage{
addIterativeLSI(
  mae,
  experiment.name = "TileMatrix500",
  embedding.name = "iterativeLSI",
  cell.depth.column = "fragments",
  col.subset = NULL,
  rank = 30,
  iterations = 2,
  first.selection = "Top",
  num.features = 25000,
  lsi.method = 2,
  cluster.method = "Seurat",
  correlation.cutoff = 0.75,
  scale.to = 10000,
  num.threads = 4,
  seed = 1,
  total.features = 5e+05,
  filter.quantile = 0.995,
  outlier.quantiles = c(0.02, 0.98),
  binarize = TRUE
)
}
\arguments{
\item{mae}{\linkS4class{MultiAssayExperiment}}

\item{experiment.name}{String containing the name of the experiment to create the embedding from and add reduced dimensions to.}

\item{embedding.name}{String containing the name of the new iterativeLSI embedding.}

\item{cell.depth.column}{String specifying the column name in the experiment colData that contains the values to use for cell depth. For arbalist created experiments, this is probably "fragments". For ArchR created experiments, this might be "nFrags".}

\item{col.subset}{Integer vector specifying the subset of the columns to include in the iterative LSI calculations.}

\item{rank}{Integer scalar specifying the rank for irlba_realized.}

\item{iterations}{Integer scalar specifying number of LSI iterations to perform.}

\item{first.selection}{String specifying either "Top" for the most accessible features or "Var" for the most variable features.}

\item{num.features}{Integer scalar specifying the number of accessible features to select when selecting the most accessible or most variable features.}

\item{lsi.method}{Number or string indicating the order of operations in the TF-IDF normalization. Possible values are: 1 or "tf-logidf", 2 or "log(tf-idf)", and 3 or "logtf-logidf".}

\item{cluster.method}{String containing cluster method. Current options: "seurat", "scran".}

\item{correlation.cutoff}{Numeric scalar specifying the cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to sequencing depth that is greater than the correlation.cutoff, it will be excluded from analysis.}

\item{scale.to}{Numeric scalar specifying the center for TF-IDF normalization.}

\item{num.threads}{Integer scalar specifying the number of threads to be used for parallel computing.}

\item{seed}{Numeric scalar to be used as the seed for random number generation. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{total.features}{Integer scalar specifying how many features to consider for use in LSI after ranking the features by the total number of insertions. These features are the only ones used throughout the variance identification and LSI.}

\item{filter.quantile}{Numeric scalar between 0 and 1 inclusive that indicates the quantile above which features should be removed based on insertion counts prior.}

\item{outlier.quantiles}{Numeric vector specifying the lower and upper quantiles for filtering cells based on the number of accessible regions.}

\item{binarize}{Logical specifying whether to binarize the matrix while creating iterative LSI reduced dimensions.}
}
\value{
\linkS4class{MultiAssayExperiment} with iterative LSI results added to the \linkS4class{SingleCellExperiment} reduced dimensions.
}
\description{
Calculate iterative LSI dimensionality reduction and add the embedding to the reduceDim field of the specified experiment.
}
\author{
Natalie Fox
}
