% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMarkerFeatures.R
\name{addMarkerFeatures}
\alias{addMarkerFeatures}
\title{Differential feature analysis}
\usage{
addMarkerFeatures(
  mae,
  experiment.name = "PeakMatrix",
  group.colname = "Clusters",
  group.experiment.name = "TileMatrix500",
  num.threads = 4,
  sampleLabels = "Sample"
)
}
\arguments{
\item{mae}{\linkS4class{MultiAssayExperiment}.}

\item{experiment.name}{String containing the experiment name for selecting the features to use for differential analysis.}

\item{group.colname}{String specifying the colname for cell groups.}

\item{group.experiment.name}{String containing the experiment name for finding the groups in the colData. If NULL, will use experiment.name.}

\item{num.threads}{Integer scalar specifying the number of threads to be used for parallel computing.}

\item{sampleLabels}{String containing experiment colData colname for sample labels.}
}
\value{
\linkS4class{MultiAssayExperiment} with DoubletScore and DoubletEnrichment columns added to the experiment colData.
}
\description{
For each of the groups specified, perform differential analysis for the specified features by comparing that group of cells to cells with similar distribution across the features from the other cell groups.
}
\author{
Natalie Fox
}
