% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSEnrichmentScore.R
\name{addTSSEnrichmentScores}
\alias{addTSSEnrichmentScores}
\title{Add TSS Enrichment Scores}
\usage{
addTSSEnrichmentScores(
  mae,
  experiment.name = "TileMatrix500",
  gene.grs = NULL,
  force = FALSE,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{mae}{A \linkS4class{MultiAssayExperiment}}

\item{experiment.name}{String specifying the experiment name to add the TSS Enrichment to the colData}

\item{gene.grs}{A \linkS4class{GRanges} specifying genes. THe start coordinate will be used as the transcription start site.}

\item{force}{Logical. If there is already a TSS Enrichment column whether to overwrite it (TRUE) or error (FALSE).}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object indicating how matrix creation should be parallelized.}
}
\description{
Calculate and add TSS enrichment scores to the MultiAssayExperiment.
}
\author{
Natalie Fox
}
