% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUMAP.R
\name{addUMAP}
\alias{addUMAP}
\title{Add UMAP transformation from MAE}
\usage{
addUMAP(
  mae,
  name.iterative.lsi = "iterativeLSI",
  name.umap = paste0(name.iterative.lsi, "_UMAP"),
  num.neighbors = 40,
  min.dist = 0.4,
  metric = "cosine",
  seed = 1,
  sample.cells = NULL,
  num.threads.transform = 1,
  outlier.quantile = 0.9,
  ...
)
}
\arguments{
\item{mae}{MultiAssayExperiment}

\item{name.iterative.lsi}{A string specifying the name of the reduced dimensions to create the UMAP from}

\item{name.umap}{A string specifying the name of the new UMAP reduced dimensions}

\item{num.neighbors}{An integer describing the number of nearest neighbors to compute a UMAP. This argument is passed to n_neighbors in uwot::umap().}

\item{min.dist}{A number specifying effective minimum distance between embedded points. See [uwot::umap]}

\item{metric}{A string specifying metric for calculating distance. ex. "cosine", "euclidean", etc. Multiple metrics can be specified. See [uwot::umap] for details.}

\item{seed}{A number used as the seed for random number generation. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{sample.cells}{A integer specifying the number of cells to subsample for UMAP creation. The cells not subsampled will be re-projected using [uwot::umap_transform] to the Embedding. Only recommended for extremely large number of cells.}

\item{num.threads.transform}{A integer specifying the number of threads for [uwot::umap_transform]}

\item{outlier.quantile}{A numeric (0 to 1) describing the distance quantile in the subsampled cels (see `sample.cells`) to use to filter poor quality re-projections.}

\item{...}{Additional parameters to pass to [uwot::umap()]}
}
\value{
A \linkS4class{MultiAssayExperiment}
}
\description{
Creating UMAP embedding from iterative LSI reduced dimensions contained with a MAE and adding the UMAP result back to the MAE
}
\author{
Natalie Fox
}
