% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeneScoreMatrix.R
\name{calculateGeneScoreMatrix}
\alias{calculateGeneScoreMatrix}
\title{Create gene score matrix}
\usage{
calculateGeneScoreMatrix(
  tile.matrix,
  gene.grs,
  tile.grs,
  extend.outside.gene = c(1000, 1e+05),
  gene.model = "exp(-abs(x)/5000) + exp(-1)",
  gene.scale.factor = 5,
  upstream = 5000,
  downstream = 0,
  BPPARAM = BiocParallel::bpparam(),
  num.threads = 4
)
}
\arguments{
\item{tile.matrix}{tile matrix count matrix.}

\item{gene.grs}{A \linkS4class{GRanges} specifying gene coordinates.}

\item{tile.grs}{A \linkS4class{GRanges} specifying tile coordinates.}

\item{extend.outside.gene}{A numeric vector with two values specifying the range to extend outside the gene prior to overlapping genes with tiles.}

\item{gene.model}{A string giving a "gene model function" used for weighting peaks for gene score calculation. This string
should be a function of `x`, where `x` is the stranded distance from the transcription start site of the gene.}

\item{gene.scale.factor}{A numeric scaling factor to weight genes based on the inverse of there length i.e. [(Scale Factor)/(Gene Length)]. This
is scaled from 1 to the scale factor. Small genes will be the scale factor while extremely large genes will be closer to 1. This scaling helps with
the relative gene score value.}

\item{upstream}{Integer describing the number of bp upstream the gene to extend the gene body. This effectively makes the gene body larger as there are proximal peaks that should be weighted equally to the gene body.}

\item{downstream}{Integer describing the number of bp downstream the gene to extend the gene body.This effectively makes the gene body larger as there are proximal peaks that should be weighted equally to the gene body.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object indicating how doublet scoring should be parallelized per iteration of doublets.}

\item{num.threads}{Integer scalar specifying the number of threads to use for tatami multiply.}
}
\value{
\linkS4class{SingleCellExperiment}
}
\description{
Calculate gene score matrix to estimate gene expression.
}
\author{
Natalie Fox
}
