% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDoubletScores.R
\name{combineDoubletResults}
\alias{combineDoubletResults}
\title{Combine colData on doublet results}
\usage{
combineDoubletResults(mae.list, experiment.name = "TileMatrix500")
}
\arguments{
\item{mae.list}{List of \linkS4class{MultiAssayExperiment}. They should be identical other than Each will be values for different DoubletScore and DoubletEnrichment colData of the specified experiment. The input MultiAssayExperiments can be created using addDoubletScores on the same MAE while specifying different samples to the selected.samples argument.}

\item{experiment.name}{String containing the experiment name to calculate doublet scores.}
}
\description{
Combine doublet result colData into the first MAE in the list.
}
\author{
Natalie Fox
}
