% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArbalistMAE.R
\name{createArbalistMAE}
\alias{createArbalistMAE}
\title{Import a single cell ATAC-seq MultiAssayExperiment}
\usage{
createArbalistMAE(
  sample.names,
  fragment.files,
  filtered.feature.matrix.files,
  barcode.annotation.files = NULL,
  sample.annotation.files = NULL,
  output.dir = tempdir(),
  multiome = NULL,
  min.frags = 1000,
  max.frags = Inf,
  tile.size = 500,
  seq.lengths = NULL,
  gene.grs = NULL,
  use.alt.exp = FALSE,
  main.exp.name = "TileMatrix500",
  filter.rna.features.without.intervals = TRUE,
  BPPARAM = bpparam()
)
}
\arguments{
\item{sample.names}{String vector containing the sample names. Must be the same length as fragment.files & filtered.feature.matrix.files.}

\item{fragment.files}{String vector containing the fragment file names and paths. ex. 10x Cell Ranger result output atac_fragments.tsv.gz or fragments.tsv.gz}

\item{filtered.feature.matrix.files}{String vector containing the filtered feature matrix file names and paths. ex. 10x Cell Ranger result output filtered_feature_bc_matrix.h5 or filtered_tf_bc_matrix.h5}

\item{barcode.annotation.files}{String vector containing barcode annotation to put in the SingleCellExperiment colData. ex. 10x Cell Ranger result output per_barcode_metrics.csv or singlecell.csv}

\item{sample.annotation.files}{String vector containing sample annotation to put in the MultiAssayExperiment colData. ex. 10x Cell Ranger result output summary.csv}

\item{output.dir}{String containing the directory where files should be output while creating the \linkS4class{MultiAssayExperiment}.}

\item{multiome}{Logical whether to use createMultiomeRNASCE on the filtered.feature.matrix.files to extract the RNA features and create a SingleCellExperiment. If NULL, then will become TRUE if filtered.feature.matrix.files contain "filtered_feature_bc_matrix.h5" otherwise FALSE.}

\item{min.frags}{Number specifying the minimum number of mapped ATAC-seq fragments required per cell to pass filtering for use in downstream analyses. Cells containing greater than or equal to min.frags total fragments will be retained.}

\item{max.frags}{Number specifying the maximum number of mapped ATAC-seq fragments required per cell to pass filtering for use in downstream analyses. Cells containing less than or equal to max.frags total fragments will be retained.}

\item{tile.size}{Integer scalar specifying the size of the tiles in base pairs.}

\item{seq.lengths}{Named integer vector containing the lengths of the reference sequences used for alignment.
Vector names should correspond to the names of the sequences, in the same order of occurrence as in the fragment file.
If \code{NULL}, this is obtained from the reference genome used by Cellranger (itself located by scanning the header of the fragment file).}

\item{gene.grs}{Genomic Ranges specifying gene coordinates for creating the gene score matrix. If NULL, then the geneset will be selected based on the genome version.}

\item{use.alt.exp}{Logical for selecting the MultiAssayExperiment structure. TRUE means that there will only be one experiment in the MultiAssayExperiment and all other experiments will be in alternative experiments. This option is only available if the columns are the same for all Matrices. FALSE means that each Matrix will be a separate experiment in the MAE.}

\item{main.exp.name}{String containing the name of the experiment that will be the main experiment when use.alt.exp is TRUE.}

\item{filter.rna.features.without.intervals}{Logical whether to remove GeneExpression Matrix features from the h5.files that do not have interval specified. Often these are mitochondria genes.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object indicating how matrix creation should be parallelized.}
}
\value{
A \linkS4class{MultiAssayExperiment}
}
\description{
Import scATAC-seq or multiome results into a MultiAssayExperiment.
}
\author{
Natalie Fox
}
