% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGenomeRefFilenameFromFragmentFile.R
\name{extractGenomeRefFilenameFromFragmentFile}
\alias{extractGenomeRefFilenameFromFragmentFile}
\title{Retrieves genome reference file name}
\usage{
extractGenomeRefFilenameFromFragmentFile(fragment.file)
}
\arguments{
\item{fragment.file}{String specifying fragment file name}
}
\value{
Named string vector
}
\description{
Helper function to extract genome reference file name from fragment file header.
}
\examples{
\dontrun{
download.file(
paste0("https://cf.10xgenomics.com/samples/cell-arc/1.0.0/",
"pbmc_granulocyte_sorted_10k/",
"pbmc_granulocyte_sorted_10k_atac_fragments.tsv.gz"), 
file.path(tempdir(), "pbmc_granulocyte_sorted_10k_atac_fragments.tsv.gz")
)
fragment.file <- file.path(tempdir(), "pbmc_granulocyte_sorted_10k_atac_fragments.tsv.gz")
extractGenomeRefFilenameFromFragmentFile(fragment.file)
}

}
\author{
Natalie Fox
}
