% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDuplicateFeatures.R
\name{filterDuplicateFeatures}
\alias{filterDuplicateFeatures}
\title{Filter duplicate features}
\usage{
filterDuplicateFeatures(
  se,
  mcol.name = "name",
  summary.stat = sum,
  selection.matrix = max
)
}
\arguments{
\item{se}{\linkS4class{SummarizedExperiment}}

\item{mcol.name}{String specifying the colname for the experiment rowData}

\item{summary.stat}{Function to summarize each feature (row) of the experiment}

\item{selection.metric}{Function to select the row to keep when there are duplicate rows with the same mcol.name}
}
\description{
Keeps one feature from each set of duplicated features based on specified criteria, such as the feature with the highest values. Other duplicate features are removed.
}
