% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{findReducedDimRes}
\alias{findReducedDimRes}
\title{Return reduced dimensions matrix with the reduced dimension name}
\usage{
findReducedDimRes(mae, name.reduced.dim)
}
\arguments{
\item{mae}{\linkS4class{MultiAssayExperiment}}

\item{name.reduced.dim}{String specifying the the reduced dimensions name to return. String can be a named vector where the name in the experiment in case there are reduced dimensions within multiple experiments with the same names.}
}
\description{
Checks through the MultiAssayExperiment both the main and alternative experiments and returns the reduced dimension with the specified name.
}
