% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{findSCE}
\alias{findSCE}
\title{Return SingleCellExperiment with the experiment name}
\usage{
findSCE(mae, experiment.name)
}
\arguments{
\item{mae}{\linkS4class{MultiAssayExperiment}}

\item{experiment.name}{String specifying the the experiment name to return. Name can either be in names() of the MAE or the altExp name in one of the experiments.}
}
\description{
Checks through the MultiAssayExperiment both the main and alternative experiments and returns the experiment with the specified name.
}
