% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUMAPFromIterativeLSI.R
\name{getUMAPFromIterativeLSI}
\alias{getUMAPFromIterativeLSI}
\title{Perform UMAP transformation}
\usage{
getUMAPFromIterativeLSI(
  x,
  num.neighbors = 40,
  min.dist = 0.4,
  metric = "cosine",
  seed = 1,
  sample.cells = NULL,
  num.threads.transform = 1,
  outlier.quantile = 0.9,
  ...
)
}
\arguments{
\item{x}{a matrix containing iterativeLSI features by cells}

\item{num.neighbors}{An number describing the number of nearest neighbors to compute a UMAP. This argument is passed to n_neighbors in uwot::umap().}

\item{min.dist}{A number specifying effective minimum distance between embedded points. See [uwot::umap]}

\item{metric}{A string specifying metric for calculating distance. ex. "cosine", "euclidean", etc. Multiple metrics can be specified. See [uwot::umap] for details.}

\item{seed}{A number used as the seed for random number generation. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{sample.cells}{A integer specifying the number of cells to subsample for UMAP creation. The cells not subsampled will be re-projected using [uwot::umap_transform] to the Embedding. Only recommended for extremely large number of cells.}

\item{num.threads.transform}{A integer specifying the number of threads for [uwot::umap_transform]}

\item{outlier.quantile}{A numeric (0 to 1) describing the distance quantile in the subsampled cels (see `sample.cells`) to use to filter poor quality re-projections.}

\item{...}{Additional parameters to pass to [uwot::umap()]}
}
\value{
A data frame containing UMAP features by cells
}
\description{
Creating UMAP matrix from the iterative LSI matrix
}
\author{
Natalie Fox
}
