% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCh38.gene.grs}
\alias{GRCh38.gene.grs}
\title{Gene features for GRCh38}
\format{
A GenomicRanges object
}
\usage{
GRCh38.gene.grs
}
\description{
GRCh38 Gene features for use in the vignette
}
\keyword{datasets}
