% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCh38.seq.lengths}
\alias{GRCh38.seq.lengths}
\title{Reference sequence lengths for GRCh38}
\format{
A names integer vector
}
\usage{
GRCh38.seq.lengths
}
\description{
GRCh38 reference sequence lengths for use in the vignette
}
\keyword{datasets}
