% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCh38.tss.grs}
\alias{GRCh38.tss.grs}
\title{TSS features for GRCh38}
\format{
A GenomicRanges object
}
\usage{
GRCh38.tss.grs
}
\description{
GRCh38 TSS features for use in the vignette
}
\keyword{datasets}
