% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCm38.tss.grs}
\alias{GRCm38.tss.grs}
\title{TSS features for GRCh38}
\format{
A GenomicRanges object
}
\usage{
GRCm38.tss.grs
}
\description{
GRCm38 TSS features for use in the vignette
}
\keyword{datasets}
