% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotFragmentSizeDistribution}
\alias{plotFragmentSizeDistribution}
\title{Plot Fragment size distribution}
\usage{
plotFragmentSizeDistribution(mae, sample.name)
}
\description{
Plot the fragment size distribution for a samples.
}
\author{
Natalie Fox
}
