% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRegionMatrix.R
\name{saveRegionMatrix}
\alias{saveRegionMatrix}
\title{Save a region count matrix}
\usage{
saveRegionMatrix(
  fragment.file,
  output.file,
  output.name,
  regions,
  barcodes = NULL,
  compress.level = 6,
  chunk.dim = 20000
)
}
\arguments{
\item{fragment.file}{String containing a path to a fragment file.}

\item{output.file}{String containing a path to an output HDF5 file.
If none exists, one will be created.}

\item{output.name}{String containing the name of the group inside \code{output.file}, in which to save the matrix contents.}

\item{regions}{A \linkS4class{GRanges} or \linkS4class{GRangesList} of the regions in which to count fragments.
If \code{regions} is a GRangesList, the count for each entry of \code{regions} will include all fragment starts/ends falling in any of the genomic intervals in its GRanges.}

\item{barcodes}{Character vector of cell barcodes to extract, e.g., based on the filtered cells reported by Cellranger.
If \code{NULL}, all barcodes are extracted, though this is usually undesirable as not all barcodes correspond to cell-containing droplets.}

\item{compress.level}{Integer scalar specifying the Zlib compression level to use.}

\item{chunk.dim}{Integer scalar specifying the size of the chunks (in terms of the number of elements) inside the HDF5 file.}
}
\value{
A sparse matrix is saved to \code{output.file} using the 10X HDF5 format.
A \linkS4class{H5SparseMatrix} referencing this file is returned where the rows correspond to entries of \code{regions}.
Column names are set to the cell barcodes - if \code{barcodes} is supplied, this is directly used as the column names.
}
\description{
Count the number of fragment start/end positions within user-defined genomic regions, using the fragment file generated by Cellranger.
Then, save the resulting matrix into a HDF5 file.
}
\details{
We count the overlap with the start/end positions of each fragment, not the overlap with the fragment interval itself.
This is because the fragment start/ends represent the transposase cleavage events, while the fragment interval has no real biological significance.

If a fragment start/end overlaps with multiple entries of \code{regions}, it is not counted as the assignment is ambiguous.
This should not be confused with overlaps to multiple intervals of the same entry when \code{regions} is a GRangesList;
this yields a mapping to a single feature, even if the exact interval isnot  ambiguous.

If the start and end for the same fragment overlap different entries of \code{regions}, the counts for both entries are incremented by 1. 
This reflects the fact that these positions represent distinct transposase cleavage events for different features.
However, if the start and end for the same fragment overlap the same entry of \code{regions}, the entry's count is only incremented by 1. 
This ensures that the count for each entry of \code{regions} still follows Poisson noise and avoids an artificial enrichment of even counts.
}
\examples{
# Mocking up the fragment file.
seq.lengths <- c(chrA = 2000, chrB = 10000)
temp <- tempfile(fileext = ".gz")
mockFragmentFile(temp, seq.lengths, 1e3, cell.names = LETTERS)

# Mocking up some regions of interest.
library(GenomicRanges)
regions <- GRanges(c("chrA:500-1000", "chrB:1000-2000"))

# Running the counter
out <- tempfile(fileext=".h5")
counts <- saveRegionMatrix(temp, output.file=out, output.name="WHEE", regions=regions)
counts

}
\author{
Aaron Lun
}
