% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSEnrichmentScore.R
\name{tssEnrichmentScores}
\alias{tssEnrichmentScores}
\title{Calculate TSS Enrichment Scores}
\usage{
tssEnrichmentScores(
  fragment.file,
  barcodes,
  gene.grs,
  window = 101,
  norm = 100,
  flank = 2000,
  min.norm = 0.2
)
}
\arguments{
\item{fragment.file}{String specifying fragment file}

\item{barcodes}{Vector or strings specified the cell barcodes to include from the fragment file}

\item{gene.grs}{GRanges specifying gene where there start coordinate will be used as the TSS}

\item{window}{Number specifying the size in bp of the TSS window}

\item{norm}{Number specifying the size in bp of the flanking region window}

\item{flank}{Number specifying the bp distance for the flanking region from the TSS}

\item{min.norm}{}
}
\description{
Calculate TSS enrichment scores from one fragment.files.
}
\author{
Natalie Fox
}
