% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDoubletScores.R
\name{addDoubletScores}
\alias{addDoubletScores}
\title{Add doublet scores}
\usage{
addDoubletScores(
  mae,
  experiment.name = "TileMatrix500",
  seed = 5,
  num.trials = 5,
  num.threads = 1,
  num.dimensions = 30,
  num.doublets.per.iteration = 200,
  max.num.synthetic.doublets = 2000,
  plot.out.dir = ".",
  selected.samples = NULL,
  BPPARAM = bpparam()
)
}
\arguments{
\item{mae}{\linkS4class{MultiAssayExperiment}.}

\item{experiment.name}{String containing thr experiment name to calculate doublet scores.}

\item{seed}{Numeric scalar to be used as the seed for random number generation. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{num.trials}{Integer scalar specifying the number of times to simulate the number of cells in the sample for doublets.}

\item{num.threads}{Integer scalar specifying the number of threads to be used for parallel computing.}

\item{num.dimensions}{Integer scalar specifying the number of iterative LSI dimensions to use doublet scoring.}

\item{num.doublets.per.iteration}{Integer scalar specifying how many doublets to estimate at one time this allows adjustment of memory usage/run time.}

\item{max.num.synthetic.doublets}{Integer scalar specifying the maximum number of doublets to run per sample. This is used if num.trials * number of cells in the sample is more than this number.}

\item{plot.out.dir}{String specifying the output path for doublet score quality control plots.}

\item{selected.samples}{Vector of strings specifying which samples to calculate doublet scores for. By default (NULL), all that have enough cells will be run.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object indicating how doublet scoring should be parallelized per iteration of doublets.}
}
\value{
\linkS4class{MultiAssayExperiment} with DoubletScore and DoubletEnrichment columns added to the experiment colData.
}
\description{
Calculate doublet scores and add them to the MultiAssayExperiment.
}
\author{
Natalie Fox
}
