% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGroupCoverages.R
\name{addGroupCoverages}
\alias{addGroupCoverages}
\title{Create Pseudobulk replicates}
\usage{
addGroupCoverages(
  mae,
  experiment.name = "TileMatrix500",
  group.by.colname = "Clusters",
  sampleLabels = "Sample",
  min.cells = 40,
  max.cells = 500,
  max.frags = 25 * 10^6,
  min.reps = 2,
  max.reps = 5,
  sample.ratio = 0.8,
  kmerLength = 6,
  return.groups = FALSE,
  force = FALSE,
  coverage.file.path = getwd(),
  BPPARAM = bpparam()
)
}
\value{
\linkS4class{MultiAssayExperiment}
}
\description{
Select cells for pseudobulk replicats, create coverage files and add a pseudobulk experiment to the MultiAssayExperiment.
}
\author{
Natalie Fox
}
