% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIterativeLSI.R
\name{addIterativeLSI}
\alias{addIterativeLSI}
\title{Add new iterative LSI embeddings to MultiAssayExperiment}
\usage{
addIterativeLSI(
  mae,
  experiment.name = "TileMatrix500",
  embedding.name = "iterativeLSI",
  rank = 30,
  iterations = 2,
  num.features = 25000,
  cluster.method = "kmeans",
  cluster.k = 20,
  correlation.cutoff = 0.75,
  scale.to = 10000,
  num.threads = 1,
  seed = 5,
  total.features = 5e+05,
  filter.quantile = 0.995
)
}
\arguments{
\item{mae}{\linkS4class{MultiAssayExperiment}}

\item{experiment.name}{String specifying the name of the experiment to create the embedding from and add reduced dimensions to}

\item{embedding.name}{String sepcifying the name of the new iterativeLSI embedding}

\item{rank}{Number specifying the rank for irlba_realized}

\item{iterations}{Number of LSI iterations to perform.}

\item{num.features}{Number of accessible features to select when selecting the most accessible or most variable features.}

\item{cluster.method}{String specifying cluster method. Currently "kmeans" is the only supported option.}

\item{cluster.k}{Number of clusters to use}

\item{correlation.cutoff}{Numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to sequencing depth that is greater than the corCutOff, it will be excluded from analysis.}

\item{scale.to}{Number specifying the center for TF-IDF normalization.}

\item{num.threads}{Number of threads to be used for parallel computing.}

\item{seed}{Number to be used as the seed for random number generation. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{total.features}{Number of features to consider for use in LSI after ranking the features by the total number of insertions. These features are the only ones used throughout the variance identification and LSI.}

\item{filter.quantile}{Number 0,1 that indicates the quantile above which features should be removed based on insertion counts prior}
}
\value{
\linkS4class{MultiAssayExperiment}
}
\description{
Add new iterative LSI embeddings to MultiAssayExperiment
}
\examples{
\dontrun{
mae <- maw.scatac::importScAtac('FRS15024')
mae.res <- addIterativeLSI(mae)
}
}
\author{
Natalie Fox
}
