% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPeakMatrix.R
\name{addPeakMatrix}
\alias{addPeakMatrix}
\title{Create a peak matrix}
\usage{
addPeakMatrix(
  mae,
  genome.size,
  pseudobulk.experiment.name = "TileMatrix500_pseudobulk",
  sc.experiment.name = "TileMatrix500",
  gene.grs = NULL,
  reproducibility = "2",
  shift = -75,
  extsize = 150,
  method = "q",
  cutOff = 0.1,
  nomodel = TRUE,
  nolambda = TRUE,
  extendSummits = 250,
  output.dir = tempdir(),
  BPPARAM = bpparam()
)
}
\arguments{
\item{mae}{\linkS4class{MultiAssayExperiment}.}

\item{genome.size}{Integer scalar specifying the effective genome size or the size of the genome that is mappable to use in MACSr peak calling. See [MACSr::callpeak] for more details.}

\item{pseudobulk.experiment.name}{String containing Experiment name for the pseudobulk \linkS4class{SummarizedExperiment} containing colData pointing to coverage files. This experiment can be created from arbalist::addGroupCoverages.}

\item{sc.experiment.name}{String containing experiment name for \linkS4class{SingleCellExperiment} with cell barcodes as the rownames.}

\item{gene.grs}{A \linkS4class{GRanges}. If the The start coordinate will be used as the transcription start site.}

\item{reproducibility}{String that indicates how peak reproducibility should be handled. This string is dynamic and can be a function of n where n is the number of samples being assessed. For example, reproducibility = "2" means at least 2 samples must have a peak call at this locus and reproducibility = "(n+1)/2" means that the majority of samples must have a peak call at this locus.}

\item{shift}{Integer scalar specifying how many base pairs to shift each Tn5 insertion. When combined with extsize this allows you to create proper fragments, centered at the Tn5 insertion site, for use with MACSr::callpeak.}

\item{extsize}{Integer scalar specifying how many base pairs to extend the fragment after shift has been applied. When combined with extsize this allows you to create proper fragments, centered at the Tn5 insertion site, for use with MACSr::callpeak.}

\item{method}{String containing the method to use for significance testing in MACS2. Options are "p" for p-value and "q" for q-value. When combined with cutOff this gives the method and significance threshold for peak calling (see MACS2 documentation).}

\item{cutOff}{Numeric scalar specifying significance cutOff for the testing method indicated by method (see MACSr::callpeak documentation).}

\item{nomodel}{Logical specifying whether or not to build the shifting model during MACSr::callpeak.}

\item{nolambda}{If True, MACS will use fixed background lambda as local lambda for every peak region.}

\item{extendSummits}{Integer scalar specifying how many base pairs to extend peak summits (in both directions) to obtain final fixed-width peaks. For example, extendSummits = 250 will create 501-bp fixed-width peaks from the 1-bp summits.}

\item{output.dir}{String containing the directory where hdf5 files should be output while creating the peak matrix \linkS4class{SingleCellExperiment}.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object indicating how matrix creation should be parallelized.}
}
\value{
\linkS4class{MultiAssayExperiment} with a new peak matrix \linkS4class{SingleCellExperiment}
}
\description{
Call peaks using MACSr and then create a peak matrix experiment in the MultiAssayExperiment.
}
\author{
Natalie Fox
}
