% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPeakMatrix.R
\name{addPeakMatrix}
\alias{addPeakMatrix}
\title{Create a peak matrix}
\usage{
addPeakMatrix(
  mae,
  genome.size,
  pseudobulk.experiment.name = "TileMatrix500_pseudobulk",
  sc.experiment.name = "TileMatrix500",
  reproducibility = "2",
  shift = -75,
  extsize = 150,
  method = "q",
  cutOff = 0.1,
  nomodel = TRUE,
  nolambda = TRUE,
  extendSummits = 250,
  output.dir = tempdir(),
  BPPARAM = bpparam(),
  ...
)
}
\arguments{
\item{mae}{\linkS4class{MultiAssayExperiment}}

\item{genome.size}{Number specifying the effective genome size or the size of hte genome that is mappable to use in MACSr peak calling. See [MACSr::callpeak] for more details.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object indicating how matrix creation should be parallelized.}
}
\value{
\linkS4class{MultiAssayExperiment}
}
\description{
Call peaks using MACSr and then create a peak matrix experiment in the MultiAssayExperiment.
}
\author{
Natalie Fox
}
