% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSEnrichmentScore.R
\name{addTSSEnrichmentScores}
\alias{addTSSEnrichmentScores}
\title{Add TSS Enrichment Scores}
\usage{
addTSSEnrichmentScores(
  mae,
  experiment.name = "TileMatrix500",
  experiment.name.for.gene.grs = "GeneExpressionMatrix",
  gene.grs = NULL,
  window = 101,
  norm = 100,
  flank = 2000,
  min.norm = 0.2,
  force = FALSE,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{mae}{A \linkS4class{MultiAssayExperiment}}

\item{experiment.name}{String specifying the experiment name to add the TSS Enrichment to the colData}

\item{experiment.name.for.gene.grs}{String specifying the experiment name to use rowRanges start as for TSS. If this experiment name is not found in mae, then gene.grs must be specified. If both experiment.name.for.gene.grs and gene.grs are specified, then gene.grs takes precendence.}

\item{gene.grs}{A \linkS4class{GRanges} specifying genes. The start coordinate will be used as the transcription start site.}

\item{window}{Number specifying the size in bp of the TSS window.}

\item{norm}{Number specifying the size in bp of the flanking region window.}

\item{flank}{Number specifying the bp distance for the flanking region from the TSS.}

\item{min.norm}{Number specifying the minimum allow for flanking region window. Avoids divide by zero error.}

\item{force}{Logical. If there is already a TSS Enrichment column whether to overwrite it (TRUE) or error (FALSE).}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object indicating how matrix creation should be parallelized.}
}
\description{
Calculate and add TSS enrichment scores to the MultiAssayExperiment.
}
\author{
Natalie Fox
}
