% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArbalistMAEFromCellrangerDirs.R
\name{createArbalistMAEFromCellrangerDirs}
\alias{createArbalistMAEFromCellrangerDirs}
\title{Import a single cell ATAC-seq MultiAssayExperiment}
\usage{
createArbalistMAEFromCellrangerDirs(
  cellranger.res.dirs,
  output.dir = tempdir(),
  min.frags = 1000,
  max.frags = Inf,
  tile.size = 500,
  seq.lengths = NULL,
  gene.grs = NULL,
  use.alt.exp = FALSE,
  main.exp.name = "TileMatrix500",
  BPPARAM = bpparam()
)
}
\arguments{
\item{cellranger.res.dirs}{Vector of strings specifying a sunrise scATAC-seq or multiome cellranger results directory. Vector names need to be sample names.}

\item{output.dir}{String containing the directory where files should be output while creating the \linkS4class{MultiAssayExperiment}.}

\item{min.frags}{Number specifying the minimum number of mapped ATAC-seq fragments required per cell to pass filtering for use in downstream analyses. Cells containing greater than or equal to min.frags total fragments will be retained.}

\item{max.frags}{Number specifying the maximum number of mapped ATAC-seq fragments required per cell to pass filtering for use in downstream analyses. Cells containing less than or equal to max.frags total fragments will be retained.}

\item{tile.size}{Integer scalar specifying the size of the tiles in base pairs.}

\item{seq.lengths}{Named integer vector containing the lengths of the reference sequences used for alignment.
Vector names should correspond to the names of the sequences, in the same order of occurrence as in the fragment file.
If \code{NULL}, this is obtained from the reference genome used by Cellranger (itself located by scanning the header of the fragment file).}

\item{gene.grs}{Genomic Ranges specifying gene coordinates for creating the gene score matrix. If NULL, then the geneset will be selected based on the genome version.}

\item{use.alt.exp}{Logical for selecting the MultiAssayExperiment structure. TRUE means that there will only be one experiment in the MultiAssayExperiment and all other experiments will be in alternative experiments. This option is only available if the columns are the same for all Matrices. FALSE means that each Matrix will be a separate experiment in the MAE.}

\item{main.exp.name}{String containing the name of the experiment that will be the main experiment when use.alt.exp is TRUE.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object indicating how matrix creation should be parallelized.}
}
\value{
A \linkS4class{MultiAssayExperiment}
}
\description{
Import results from scATAC-seq or multiome cellranger directories into a MultiAssayExperiment.
}
\author{
Natalie Fox
}
