% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterativeLSI.R
\name{iterativeLSI}
\alias{iterativeLSI}
\title{Calculate iterative LSI dimensionality reduction}
\usage{
iterativeLSI(
  x,
  cell.names,
  sample.names,
  cell.depths,
  col.subset = NULL,
  rank = 30,
  iterations = 2,
  first.selection = "Top",
  num.features = 25000,
  lsi.method = 2,
  cluster.method = "Seurat",
  correlation.cutoff = 0.75,
  scale.to = 10000,
  num.threads = 4,
  seed = 1,
  total.features = 5e+05,
  filter.quantile = 0.995,
  outlier.quantiles = c(0.02, 0.98),
  binarize = TRUE,
  num.cells.to.sample = 10000
)
}
\arguments{
\item{x}{input matrix for creating iterative LSI embeddings, assumes ATAC-seq so chooses top accessible features.}

\item{cell.names}{String vector specifying cell names. Length should match the number of columns in x.}

\item{sample.names}{String vector specifying sample names. Length should match the number of columns in x.}

\item{cell.depths}{String vector specifying cell depths. Length should match the number of columns in x.}

\item{col.subset}{Integer vector specifying the subset of the columns to include in the iterative LSI calculations.}

\item{rank}{Integer scalar specifying the rank for irlba_realized.}

\item{iterations}{Integer scalar specifying number of LSI iterations to perform.}

\item{first.selection}{String specifying either "Top" for the most accessible features or "Var" for the most variable features.}

\item{num.features}{Integer scalar specifying the number of accessible features to select when selecting the most accessible or most variable features.}

\item{lsi.method}{Number or string indicating the order of operations in the TF-IDF normalization. Possible values are: 1 or "tf-logidf", 2 or "log(tf-idf)", and 3 or "logtf-logidf".}

\item{cluster.method}{String containing cluster method. Current options: "seurat", "scran".}

\item{correlation.cutoff}{Numeric scalar specifying the cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to sequencing depth that is greater than the correlation.cutoff, it will be excluded from analysis.}

\item{scale.to}{Numeric scalar specifying the center for TF-IDF normalization.}

\item{num.threads}{Integer scalar specifying the number of threads to be used for parallel computing.}

\item{seed}{Numeric scalar to be used as the seed for random number generation. It is recommended to keep track of the seed used so that you can reproduce results downstream.}

\item{total.features}{Integer scalar specifying how many features to consider for use in LSI after ranking the features by the total number of insertions. These features are the only ones used throughout the variance identification and LSI.}

\item{filter.quantile}{Numeric scalar between 0 and 1 inclusive that indicates the quantile above which features should be removed based on insertion counts prior.}

\item{outlier.quantiles}{Numeric vector specifying the lower and upper quantiles for filtering cells based on the number of accessible regions.}

\item{binarize}{Logical specifying whether to binarize the matrix while creating iterative LSI reduced dimensions.}

\item{num.cells.to.sample}{Scalar integer specifying the number of cells to sample iterations prior to the last in order to perform sub-sampled LSI and sub-sampled clustering.}
}
\value{
A list is returned containing:
\itemize{
\item \code{embedding}, a matrix containing the iterativeLSI embedding
\item \code{projection}, 
\item \code{subset}, a vector of indices specifying the selected subset of features
}
}
\description{
Calculate iterative LSI dimensionality reduction
}
\author{
Natalie Fox
}
