% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotClusterConfusionMap}
\alias{plotClusterConfusionMap}
\title{Plot Cluster Confusion Matrix}
\usage{
plotClusterConfusionMap(mae)
}
\arguments{
\item{mae}{\linkS4class{MultiAssayExperiment} containing a "TileMatrix500" experiment with "Cluster" and "Sample" columns in the colData}
}
\description{
Plot the association between slusters and samples using a confusion matrix.
}
\author{
Natalie Fox
}
