% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotFragmentSizeDistribution}
\alias{plotFragmentSizeDistribution}
\title{Plot Fragment size distribution}
\usage{
plotFragmentSizeDistribution(mae, sample.name, xlim = c(0, 700))
}
\arguments{
\item{mae}{\linkS4class{MultiAssayExperiment} containing fragment_file col in the colData.}

\item{sample.name}{String specifying the sample name to plot.}

\item{xlim}{Numeric vector with a min and max for the x-axis limits.}
}
\description{
Plot the fragment size distribution for a samples.
}
\author{
Natalie Fox
}
