% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotTSSenrichmentVsNumFragments}
\alias{plotTSSenrichmentVsNumFragments}
\title{Create per sample TSS enrichment plot}
\usage{
plotTSSenrichmentVsNumFragments(mae, sample.name)
}
\arguments{
\item{mae}{\linkS4class{MultiAssayExperiment} containing a "TileMatrix500" experiment with "TSSEnrichment" and "fragments" columns in the colData}

\item{sample.name}{String speciying the sample name to plot}
}
\description{
Create a per sample quality control plot looking at TSS Enrichment compared to number unique fragments.
}
\author{
Natalie Fox
}
