% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRegionMatrix.R
\name{saveRegionMatrix}
\alias{saveRegionMatrix}
\title{Save a region count matrix}
\usage{
saveRegionMatrix(
  fragment.file,
  output.file,
  output.name,
  regions,
  barcodes = NULL,
  compress.level = 6,
  chunk.dim = 20000
)
}
\arguments{
\item{fragment.file}{String containing a path to a fragment file.}

\item{output.file}{String containing a path to an output HDF5 file.
If none exists, one will be created.}

\item{output.name}{String containing the name of the group inside \code{output.file}, in which to save the matrix contents.}

\item{regions}{A \linkS4class{GRanges} or \linkS4class{GRangesList} of the regions in which to count fragments.
For a GRangesList, each GRanges is treated as a separate feature and will count all fragment starts/ends falling in any of its intervals.}

\item{barcodes}{Character vector of cell barcodes to extract, e.g., based on the filtered cells reported by Cellranger.
If \code{NULL}, all barcodes are extracted, though this is usually undesirable as not all barcodes correspond to cell-containing droplets.}

\item{compress.level}{Integer scalar specifying the Zlib compression level to use.}

\item{chunk.dim}{Integer scalar specifying the size of the chunks (in terms of the number of elements) inside the HDF5 file.}
}
\value{
A sparse matrix is saved to \code{output.file} using the 10X HDF5 format.
A \linkS4class{H5SparseMatrix} referencing this file is returned where the rows correspond to entries of \code{regions}.
Column names are set to the cell barcodes - if \code{barcodes} is supplied, this is directly used as the column names.
}
\description{
Count the number of fragment start/end positions within user-defined genomic regions, using the fragment file generated by Cellranger.
Then, save the resulting matrix into a HDF5 file.
}
\author{
Aaron Lun
}
