% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRegionMatrix.R
\name{saveRegionMatrix}
\alias{saveRegionMatrix}
\title{Save a region count matrix}
\usage{
saveRegionMatrix(
  fragment.file,
  output.file,
  output.name,
  regions,
  barcodes = NULL,
  compress.level = 6,
  chunk.dim = 20000
)
}
\arguments{
\item{fragment.file}{String containing a path to a fragment file.}

\item{output.file}{String containing a path to an output HDF5 file.
If none exists, one will be created.}

\item{output.name}{String containing the name of the group inside \code{output.file}, in which to save the matrix contents.}

\item{regions}{A \linkS4class{GRangesList} of the regions in which to}

\item{barcodes}{Character vector of cell barcodes to extract, e.g., based on the filtered cells reported by Cellranger.
If \code{NULL}, all barcodes are extracted, though this is usually undesirable as not all barcodes correspond to cell-containing droplets.}

\item{compress.level}{Integer scalar specifying the Zlib compression level to use.}

\item{chunk.dim}{Integer scalar specifying the size of the chunks (in terms of the number of elements) inside the HDF5 file.}

\item{tile.size}{Integer scalar specifying the size of the tiles in base pairs.}
}
\value{
A sparse matrix is saved to \code{output.file} using the 10X HDF5 format.
If \code{cell.names=NULL}, a character vector is invisibly returned, containing the cell barcodes corresponding to the columns of the matrix.
Otherwise, \code{NULL} is invisibly returned.
}
\description{
Count the number of fragment start/end positions within user-defined genomic regions, using the fragment file generated by Cellranger.
Then, save the resulting matrix into a HDF5 file.
}
\author{
Aaron Lun
}
